Move function comparison methods to separate module
